Automatically enable offline mode with poor internet

+++ ios/Gorilla Groove/view-controllers/TrackContextMenu.swift
@@ -33,9 +33,7 @@ class TrackContextMenu {
                 let vc = UINavigationController(rootViewController: usersController)
                 parentVc.present(vc, animated: true)
             }))
-        }
-        
-        if view == .MY_LIBRARY {
+            
             alert.addAction(UIAlertAction(title: "Add to Playlist", style: .default, handler: { _ in
                 let playlistsController = SelectPlaylistsController(track)
                 playlistsController.modalPresentationStyle = .pageSheet
@@ -43,6 +41,9 @@ class TrackContextMenu {
                 let vc = UINavigationController(rootViewController: playlistsController)
                 parentVc.present(vc, animated: true)
             }))
+        }
+        
+        if view == .MY_LIBRARY {
             alert.addAction(UIAlertAction(title: "Delete", style: .destructive, handler: { _ in
                 ViewUtil.showAlert(message: "Delete \(track.name)?", yesText: "Delete", yesStyle: .destructive, dismissText: "Cancel") {
                     TrackService.deleteTrack(track)
